requirements.txt
python-telegram-bot==13.10
SQLAlchemy==1.4.29
psycopg2==2.9.3
ordered_set
oauth2client
gspread
pandas
matplotlib
numpy
